test: move checkpoint drift coverage
