+++ tests/test_runtime_orchestrator.py
@@ -4,11 +4,7 @@ from pathlib import Path
 import pytest
 from jsonschema import Draft202012Validator
 
-from forgeflow_runtime.orchestrator import (
-    RuntimeViolation,
-    load_runtime_policy,
-    run_route,
-)
+from forgeflow_runtime.orchestrator import RuntimeViolation, load_runtime_policy, run_route
 
 
 ROOT = Path(__file__).resolve().parents[1]
@@ -25,113 +21,6 @@ def _load_schema(name: str) -> dict:
 def _assert_schema_valid(name: str, payload: dict) -> None:
     errors = sorted(Draft202012Validator(_load_schema(name)).iter_errors(payload), key=lambda err: list(err.path))
     assert not errors, [f"{list(err.path)}: {err.message}" for err in errors]
-
-
-def _make_task_dir(tmp_path: Path) -> Path:
-    task_dir = tmp_path / "task"
-    task_dir.mkdir()
-    _write_json(
-        task_dir / "brief.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "objective": "Run a small route",
-            "in_scope": ["runtime"],
-            "out_of_scope": [],
-            "constraints": ["local only"],
-            "acceptance_criteria": ["route works"],
-            "risk_level": "low",
-        },
-    )
-    _write_json(
-        task_dir / "run-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "current_stage": "clarify",
-            "status": "in_progress",
-            "completed_gates": ["clarification_complete"],
-            "failed_gates": [],
-            "retries": {},
-            "current_task_id": "",
-            "spec_review_approved": False,
-            "quality_review_approved": False,
-        },
-    )
-    return task_dir
-
-
-def _write_medium_plan_artifacts(task_dir: Path, *, route_name: str = "medium") -> None:
-    _write_json(
-        task_dir / "plan.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "steps": [
-                {
-                    "id": "step-1",
-                    "objective": "update workflow docs",
-                    "dependencies": [],
-                    "expected_output": "workflow docs reflect medium route behavior",
-                    "verification": "pytest tests/test_runtime_orchestrator.py -q",
-                    "rollback_note": "remove incomplete workflow edits if validation fails",
-                }
-            ],
-        },
-    )
-    _write_json(
-        task_dir / "plan-ledger.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": route_name,
-            "completed_stages": [],
-            "completed_gates": [],
-            "retries": {},
-            "current_task_id": "task-1",
-            "tasks": [
-                {
-                    "id": "task-1",
-                    "title": "update workflow docs",
-                    "depends_on": [],
-                    "files": ["docs/workflow.md"],
-                    "parallel_safe": False,
-                    "status": "in_progress",
-                    "required_gates": ["machine", "validator"],
-                    "evidence_refs": [],
-                    "attempt_count": 0,
-                }
-            ],
-        },
-    )
-
-
-def test_run_route_rejects_medium_plan_ledger_out_of_order_completed_stages(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_medium_plan_artifacts(task_dir, route_name="medium")
-    run_state = json.loads((task_dir / "run-state.json").read_text(encoding="utf-8"))
-    run_state["current_stage"] = "execute"
-    _write_json(task_dir / "run-state.json", run_state)
-    plan_ledger = json.loads((task_dir / "plan-ledger.json").read_text(encoding="utf-8"))
-    plan_ledger["completed_stages"] = ["clarify", "plan", "quality-review"]
-    plan_ledger["completed_gates"] = ["clarification_complete", "plan_executable"]
-    _write_json(task_dir / "plan-ledger.json", plan_ledger)
-    _write_json(
-        task_dir / "review-report.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "review_type": "quality",
-            "verdict": "approved",
-            "findings": ["looks fine"],
-            "approved_by": "quality-reviewer",
-            "next_action": "finalize 가능",
-        },
-    )
-
-    with pytest.raises(RuntimeViolation, match="plan-ledger checkpoint has out-of-sequence completed stages at execute: quality-review"):
-        run_route(task_dir=task_dir, policy=policy, route_name="medium")
 
 
 def test_small_route_runs_end_to_end_and_updates_state(tmp_path: Path) -> None:
